Moved enumerations to types.py
